Move CorsBuilder and Session to the Http package.
This co-locates the entire Http stack together making a future sub-split
much easier.

Refs #11385
