--- src/Network/Request.php
@@ -1,4 +1,0 @@
-<?php
-// @deprecated Load new class and alias
-class_exists('Cake\Http\ServerRequest');
-deprecationWarning('Use Cake\Http\ServerRequest instead of Cake\Network\Request.');
