+++ tests/TestCase/Http/ResponseTest.php
@@ -18,10 +18,10 @@ include_once CORE_TEST_CASES . DS . 'Http' . DS . 'server_mocks.php';
 
 use Cake\Http\Cookie\Cookie;
 use Cake\Http\Cookie\CookieCollection;
+use Cake\Http\CorsBuilder;
 use Cake\Http\Exception\NotFoundException;
 use Cake\Http\Response;
 use Cake\Http\ServerRequest;
-use Cake\Network\CorsBuilder;
 use Cake\TestSuite\TestCase;
 use Zend\Diactoros\Stream;
