+++ tests/TestCase/Http/ServerRequestTest.php
@@ -20,7 +20,7 @@ use Cake\Http\Cookie\CookieCollection;
 use Cake\Http\Exception\MethodNotAllowedException;
 use Cake\Http\ServerRequest;
 use Cake\Http\ServerRequestFactory;
-use Cake\Network\Session;
+use Cake\Http\Session;
 use Cake\TestSuite\TestCase;
 use Zend\Diactoros\UploadedFile;
 use Zend\Diactoros\Uri;
